# Optimize performance of `Measurement` constructor