@stevejgordon 2024-06-27
My pleasure, @tarekgh!